requirements.txt
streamlit
pdf2image
Pillow
python-docx
python-pptx
